backend/requirements.txt
alembic==1.15.2
annotated-types==0.7.0
anyio==4.9.0
bcrypt==4.3.0
certifi==2025.4.26
cffi==1.17.1
click==8.1.8
colorama==0.4.6
contourpy==1.3.2
cryptography==44.0.2
cycler==0.12.1
distro==1.9.0
easyocr==1.7.2
ecdsa==0.19.1
fastapi==0.115.12
filelock==3.18.0
fonttools==4.57.0
fsspec==2025.3.2
greenlet==3.2.1
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.10
imageio==2.37.0
iniconfig==2.1.0
Jinja2==3.1.6
jiter==0.9.0
kiwisolver==1.4.8
lazy_loader==0.4
Mako==1.3.10
MarkupSafe==3.0.2
matplotlib==3.10.1
mpmath==1.3.0
networkx==3.4.2
ninja==1.11.1.4
numpy==2.2.5
openai==1.76.0
opencv-python-headless==4.11.0.86
packaging==25.0
passlib==1.7.4
pillow==11.2.1
pluggy==1.5.0
psycopg2-binary==2.9.10
pyasn1==0.4.8
pyclipper==1.3.0.post6
pycparser==2.22
pydantic==2.11.3
pydantic-settings==2.9.1
pydantic_core==2.33.1
pyparsing==3.2.3
pytest==8.3.5
pytest-asyncio==0.26.0
python-bidi==0.6.6
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
python-jose==3.4.0
python-multipart==0.0.20
PyYAML==6.0.2
rsa==4.9.1
scikit-image==0.25.2
scipy==1.15.2
setuptools==79.0.1
shapely==2.1.0
six==1.17.0
sniffio==1.3.1
SQLAlchemy==2.0.40
starlette==0.46.2
sympy==1.13.3
tifffile==2025.3.30
torch==2.7.0
torchvision==0.22.0
tqdm==4.67.1
typing-inspection==0.4.0
typing_extensions==4.13.2
uvicorn==0.34.2
